SERVER-14538 _appendUserInfo: don't save addr pointing to temporary
Before de724781, the return value of CurOp::getNS() had a lifetime of
the CurOp object.   As of de724781, CurOp::getNS() returns a
temporary.

_appendUserInfo() initialized a StringData from the result of
CurOp::getNS(), which was poor style but technically valid before
de724781.  As of de724781, the buffer being pointed to by the
StringData is destroyed along with the temporary, creating a
use-after-free at introspect.cpp:70 when the StringData is read.